+++ docs/Changelog.md
@@ -10,6 +10,7 @@
 - House numbers (nodes) fix visibility [#168](https://github.com/mapsforge/vtm/issues/168)
 - Use baseline 160dpi in scaling [#236](https://github.com/mapsforge/vtm/issues/236)
 - MapFileTileSource zoom level API enhancements [#219](https://github.com/mapsforge/vtm/issues/219)
+- Animator enhancements with easing functions [#246](https://github.com/mapsforge/vtm/issues/246)
 - Tile grid layer scaling [#238](https://github.com/mapsforge/vtm/issues/238)
 - Internal render themes various improvements [#41](https://github.com/mapsforge/vtm/issues/41)
 - Many other minor improvements and bug fixes